PYIC-7555 error handlers to TS
