+++ package.json
@@ -37,9 +37,11 @@
     "@types/chai": "^4.3.20",
     "@types/chai-as-promised": "^7.1.8",
     "@types/cookie-parser": "^1.4.7",
+    "@types/csurf": "^1.11.5",
     "@types/express": "^5.0.0",
     "@types/express-session": "^1.18.0",
     "@types/mocha": "^10.0.8",
+    "@types/proxyquire": "^1.3.31",
     "@types/sinon": "^17.0.3",
     "@types/sinon-chai": "^3.2.12",
     "@types/uid-safe": "^2.1.5",
