+++ src/app.ts
@@ -49,6 +49,8 @@ declare global {
 declare module "express-session" {
   interface SessionData {
     ipvSessionId: string;
+    clientOauthSessionId?: string;
+    currentPage?: string;
   }
 }
